Tests: patch requests.get per-test for FHIR proxy\n\n- Each test now patches requests.get with a side_effect that returns a fake CapabilityStatement for /metadata and a test-specific response for resource fetches\n- Removes global patching and monkeypatching of capability_index\n- All tests are now robust, isolated, and pass without a live FHIR server\n- Next: refactor test setup to reduce code repetition via fixtures/helpers
